backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic==2.9.0
websockets==13.0
xgboost==2.1.0
lightgbm==4.5.0
scikit-learn==1.5.0
imbalanced-learn==0.12.0
pyod==2.0.0
pandas==2.2.0
numpy==1.26.4
joblib==1.4.0
anthropic==0.39.0
httpx==0.27.2
python-dotenv==1.0.0
